Move node ext and parse result to their own files
